Setup window surface and present queue
I'm sticking with it to prevent my code from deviating from the tutorial too much for now, but I'm liking the use of std::optional in QueueFamilyIndices less and less. Even now that we've added a second queue family to the mix, it feels over engineered and like a mis-application of DRY to make one function do both, and return one struct with both queue families in it.